backend/requirements.txt
flask
flask-restful
flask-sqlalchemy
numpy
flask-cors
sqlalchemy
datetime
python-dotenv
flask-security
argon2_cffi